+++ src/SmartHopper.Components/Text/AITextGenerate.cs
@@ -194,7 +194,7 @@ namespace SmartHopper.Components.Text
                 }
 
                 this.parent.SetPersistentOutput("Result", value, DA);
-                message = "Done :)";
+                message = string.Empty;
             }
         }
     }