+++ src/SmartHopper.Infrastructure/AIProviders/AIProvider.cs
@@ -38,7 +38,7 @@ namespace SmartHopper.Infrastructure.AIProviders
     /// <typeparam name="T">The type of the derived provider class.</typeparam>
     public abstract class AIProvider<T> : AIProvider where T : AIProvider<T>
     {
-        private static readonly Lazy<T> InstanceValue = new (() => Activator.CreateInstance(typeof(T), true) as T);
+        private static readonly Lazy<T> InstanceValue = new(() => Activator.CreateInstance(typeof(T), true) as T);
 
         /// <summary>
         /// Initializes a new instance of the <see cref="AIProvider{T}"/> class.